requirements.txt
mcp[cli]
openai
fastapi
python-dotenv
loguru # or use logging
requests
pytz
